fetch data only once

+++ tides-app/src/features/station/fetch.js
@@ -1,11 +1,9 @@
-/* global BigInt */
-
 import fetchJson from "fetch-json";
 
 const BASE_URL = "https://www.pegelonline.wsv.de/webservices/rest-api/v2/stations";
 
 function getUnixTimestamp(date) {
-    return BigInt(date.getTime() / 1000);
+    return date.getTime() / 1000;
 }
 
 export default async function fetchMeasurements(stationId) {
